Moving enums to its own file. Refactoring code to get ready for the big one
