--- gaia/buildings.py
@@ -1,9 +1,0 @@
-from enum import IntEnum
-
-
-class Building(IntEnum):
-    MINE = 0
-    TRADING_STATION = 1
-    RESEARCH_LAB = 2
-    PLANETARY_INSTITUTE = 3
-    ACADEMY = 4
