+++ tests/test_map.py
@@ -2,10 +2,8 @@ import pytest
 from copy import copy, deepcopy
 import json
 
-from gaia.map import Hexagon, Planet, InhabitedPlanet, Sector, Map, GameTile
-from gaia.players import Factions
-from gaia.buildings import Building
-from gaia.planet_types import PlanetType
+from gaia.board.map import Hexagon, Planet, InhabitedPlanet, Sector, Map, GameTile
+from gaia.enums import PlanetType, Factions, Building
 
 
 @pytest.mark.parametrize("hex1,hex2,distance", [
